requirements.txt
asgiref==3.5.0
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
cloudpickle==2.0.0
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
dask==2022.2.1
Django==4.0.3
django-eventstream==4.4.0
django-filter==21.1
django-grip==3.0.0
djangorestframework==3.13.1
fsspec==2022.2.0
gripcontrol==4.1.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.2
incremental==21.3.0
locket==0.2.1
Markdown==3.3.6
packaging==21.3
partd==1.2.0
pubcontrol==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytz==2021.3
PyYAML==6.0
requests==2.27.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toolz==0.11.2
Twisted==22.2.0
txaio==22.2.1
typing-extensions==4.1.1
urllib3==1.26.8
Werkzeug==0.16.1
zipp==3.7.0
zope.interface==5.4.0
